Fix: Retain a consistent `mrb_gc.arena` state
The `mrb_gc.arena` is a pool of object slots used for temporarily keeping objects alive while they're being manipulated in C and are otherwise unreferenced. By swapping out — and resetting — this pool, we were previously causing unpredictable harm to the state of the mruby runtime.

Considering we never really had a reason to do this, it seems perfectly reasonable to fallback to reusing the original pool.